+++ util/io.py
@@ -28,3 +28,11 @@ warnings.filterwarnings ("once", message=str(mcpdft_removal_warning),
 def mcpdft_removal_warn (): warnings.warn (mcpdft_removal_warning,
                                            stacklevel=3)
 
+lassi_dir_warning = FutureWarning((
+    "LASSI code is being moved from my_pyscf.mcscf.lassi_* module files "
+    "to my_pyscf.lassi.* module files; the former will be removed soon."
+    ))
+warnings.filterwarnings ("once", message=str(lassi_dir_warning),
+                         category=FutureWarning)
+def lassi_dir_warn (): warnings.warn (lassi_dir_warning, stacklevel=3)
+
